INTERNAL: set의 get, delete 로직 hash tree로 추출
