+++ engines/default/coll_set.c
@@ -51,52 +51,6 @@ static inline void UNLOCK_CACHE(void)
 }
 
 /*
- * Hash table management
- */
-typedef struct hash_node {
-    int cnt;
-    int keys[3];
-} hash_node;
-
-typedef struct hash_table {
-    hash_node *buckets;
-    int capacity;
-} hash_table;
-
-static bool hash_init(hash_table *ht, int count)
-{
-    ht->capacity = (int)(1.3 * (double)count);
-    ht->buckets = (hash_node*)calloc(ht->capacity, sizeof(hash_node));
-    if (ht->buckets == NULL)
-        return false;
-    return true;
-}
-
-static void hash_free(hash_table *ht)
-{
-    if (ht->buckets) {
-        free(ht->buckets);
-        ht->buckets = NULL;
-        ht->capacity = 0;
-    }
-}
-
-static bool hash_insert(hash_table *ht, int key)
-{
-    size_t idx = key % ht->capacity;
-    hash_node *bucket = &ht->buckets[idx];
-    if (bucket->cnt == 3)
-        return false;
-    for (int i = 0; i < bucket->cnt; i++) {
-        if (bucket->keys[i] == key)
-            return false;
-    }
-    bucket->keys[bucket->cnt] = key;
-    bucket->cnt++;
-    return true;
-}
-
-/*
  * SET collection manangement
  */
 static const void *set_get_key(const htree_elem_item *elem, uint16_t *nkey)
@@ -107,18 +61,6 @@ static const void *set_get_key(const htree_elem_item *elem, uint16_t *nkey)
 }
 
 static htree_ops set_htree_ops = { set_get_key };
-
-typedef struct _set_prev_info {
-    htree_node *node;
-    set_elem_item *prev;
-    uint16_t       hidx;
-} set_prev_info;
-
-static inline int set_hash_eq(const int h1, const void *v1, size_t vlen1,
-                              const int h2, const void *v2, size_t vlen2)
-{
-    return (h1 == h2 && vlen1 == vlen2 && memcmp(v1, v2, vlen1) == 0);
-}
 
 static inline uint32_t do_set_elem_ntotal(set_elem_item *elem)
 {
@@ -189,11 +131,6 @@ static hash_item *do_set_item_alloc(const void *key, const uint32_t nkey,
     return it;
 }
 
-static void do_set_node_free(htree_node *node)
-{
-    do_item_mem_free(node, sizeof(htree_node));
-}
-
 static set_elem_item *do_set_elem_alloc(const uint32_t nbytes)
 {
     size_t ntotal = sizeof(set_elem_item) + nbytes;
@@ -228,60 +165,6 @@ static void do_set_elem_release(set_elem_item *elem)
     }
 }
 
-static void do_set_node_unlink(htree_meta *htree,
-                               htree_node *par_node, const int par_hidx)
-{
-    htree_node *node;
-
-    if (par_node == NULL) {
-        node = htree->root;
-        htree->root = NULL;
-        assert(node->tot_elem_cnt == 0);
-    } else {
-        assert(par_node->hcnt[par_hidx] == -1); /* child hash node */
-        set_elem_item *head = NULL;
-        set_elem_item *elem;
-        int hidx, fcnt = 0;
-
-        node = (htree_node *)par_node->htab[par_hidx];
-        assert(node->tot_elem_cnt < (HTREE_MAX_HASHCHAIN_SIZE/2));
-
-        for (hidx = 0; hidx < HTREE_HASHTAB_SIZE; hidx++) {
-            assert(node->hcnt[hidx] >= 0);
-            if (node->hcnt[hidx] > 0) {
-                fcnt += node->hcnt[hidx];
-                while (node->htab[hidx] != NULL) {
-                    elem = node->htab[hidx];
-                    node->htab[hidx] = elem->next;
-                    node->hcnt[hidx] -= 1;
-
-                    elem->next = head;
-                    head = elem;
-                }
-                assert(node->hcnt[hidx] == 0);
-            }
-        }
-        assert(fcnt == node->tot_elem_cnt);
-        node->tot_elem_cnt = 0;
-
-        par_node->htab[par_hidx] = head;
-        par_node->hcnt[par_hidx] = fcnt;
-    }
-
-    /* free the node */
-    do_set_node_free(node);
-}
-
-static void do_set_elem_unlink(htree_node *node, const int hidx,
-                               set_elem_item *prev, set_elem_item *elem)
-{
-    if (prev != NULL) prev->next = elem->next;
-    else              node->htab[hidx] = elem->next;
-    elem->linked--;
-    node->hcnt[hidx] -= 1;
-    node->tot_elem_cnt -= 1;
-}
-
 static void do_set_elem_delete_post(set_meta_info *info,
                                     set_elem_item *elem,
                                     enum elem_delete_cause cause)
@@ -298,41 +181,6 @@ static void do_set_elem_delete_post(set_meta_info *info,
     }
 }
 
-static set_elem_item *do_set_elem_delete_by_value(htree_meta *htree, htree_node *node,
-                                                   const int hval, const char *val, const int vlen,
-                                                   size_t *space_decreased)
-{
-    set_elem_item *deleted = NULL;
-    int hidx = HTREE_GET_HASHIDX(hval, node->hdepth);
-
-    if (node->hcnt[hidx] == -1) {
-        htree_node *child_node = node->htab[hidx];
-        deleted = do_set_elem_delete_by_value(htree, child_node, hval, val, vlen, space_decreased);
-        if (deleted) {
-            if (child_node->tot_elem_cnt < (HTREE_MAX_HASHCHAIN_SIZE/2)) {
-                do_set_node_unlink(htree, node, hidx);
-                *space_decreased += slabs_space_size(sizeof(htree_node));
-            }
-            node->tot_elem_cnt -= 1;
-        }
-    } else {
-        if (node->hcnt[hidx] > 0) {
-            set_elem_item *prev = NULL;
-            set_elem_item *elem = node->htab[hidx];
-            while (elem != NULL) {
-                if (set_hash_eq(hval, val, vlen, elem->hval, elem->value, elem->nbytes)) {
-                    do_set_elem_unlink(node, hidx, prev, elem);
-                    deleted = elem;
-                    break;
-                }
-                prev = elem;
-                elem = elem->next;
-            }
-        }
-    }
-    return deleted;
-}
-
 static ENGINE_ERROR_CODE do_set_elem_delete(set_meta_info *info,
                                             const char *val, const int vlen)
 {
@@ -341,13 +189,9 @@ static ENGINE_ERROR_CODE do_set_elem_delete(set_meta_info *info,
     if (htree->root != NULL) {
         int hval = genhash_string_hash(val, vlen);
         size_t space_decreased = 0;
-        set_elem_item *deleted = do_set_elem_delete_by_value(htree, htree->root, hval, val, vlen, &space_decreased);
+        htree_elem_item *deleted = htree_elem_delete(htree, htree->root, hval, val, vlen, &space_decreased);
         if (deleted != NULL) {
-            do_set_elem_delete_post(info, deleted, ELEM_DELETE_NORMAL);
-            if (htree->root->tot_elem_cnt == 0) {
-                do_set_node_unlink(htree, NULL, 0);
-                space_decreased += slabs_space_size(sizeof(htree_node));
-            }
+            do_set_elem_delete_post(info, (set_elem_item *)deleted, ELEM_DELETE_NORMAL);
             if (info->stotal > 0 && space_decreased > 0) {
                 do_coll_space_decr((coll_meta_info *)info, ITEM_TYPE_SET, space_decreased);
             }
@@ -355,195 +199,6 @@ static ENGINE_ERROR_CODE do_set_elem_delete(set_meta_info *info,
         }
     }
     return ret;
-}
-
-static int do_set_elem_get_all(htree_meta *htree, htree_node *node,
-                               const bool delete, set_elem_item **elem_array,
-                               size_t *space_decreased)
-{
-    assert(elem_array != NULL);
-    int hidx;
-    int fcnt = 0; /* found count */
-
-    for (hidx = 0; hidx < HTREE_HASHTAB_SIZE; hidx++) {
-        if (node->hcnt[hidx] == -1) {
-            htree_node *child_node = (htree_node *)node->htab[hidx];
-            int ecnt = do_set_elem_get_all(htree, child_node, delete, &elem_array[fcnt], space_decreased);
-            fcnt += ecnt;
-            if (delete) {
-                if (child_node->tot_elem_cnt < (HTREE_MAX_HASHCHAIN_SIZE/2)) {
-                    do_set_node_unlink(htree, node, hidx);
-                    *space_decreased += slabs_space_size(sizeof(htree_node));
-                }
-                node->tot_elem_cnt -= ecnt;
-            }
-        } else if (node->hcnt[hidx] > 0) {
-            set_elem_item *elem = node->htab[hidx];
-            while (elem != NULL) {
-                elem->refcount++;
-                elem_array[fcnt] = elem;
-                fcnt++;
-                if (delete) do_set_elem_unlink(node, hidx, NULL, elem);
-                elem = (delete ? node->htab[hidx] : elem->next);
-            }
-        }
-    }
-    return fcnt;
-}
-
-static int do_set_elem_delete_by_count(htree_meta *htree, htree_node *node,
-                                       const uint32_t count, set_elem_item **deleted_head,
-                                       size_t *space_decreased)
-{
-    int hidx;
-    int fcnt = 0;
-
-    for (hidx = 0; hidx < HTREE_HASHTAB_SIZE; hidx++) {
-        if (node->hcnt[hidx] == -1) {
-            htree_node *child_node = (htree_node *)node->htab[hidx];
-            int rcnt = (count > 0 ? (count - fcnt) : 0);
-            int ecnt = do_set_elem_delete_by_count(htree, child_node, rcnt, deleted_head, space_decreased);
-            fcnt += ecnt;
-            if (child_node->tot_elem_cnt < (HTREE_MAX_HASHCHAIN_SIZE/2)) {
-                do_set_node_unlink(htree, node, hidx);
-                *space_decreased += slabs_space_size(sizeof(htree_node));
-            }
-            node->tot_elem_cnt -= ecnt;
-        } else if (node->hcnt[hidx] > 0) {
-            set_elem_item *elem = node->htab[hidx];
-            while (elem != NULL) {
-                fcnt++;
-                do_set_elem_unlink(node, hidx, NULL, elem);
-
-                /* link deleted elem into list; returned to caller */
-                elem->next = *deleted_head;
-                *deleted_head = elem;
-
-                if (count > 0 && fcnt >= count) break;
-                elem = node->htab[hidx];
-            }
-        }
-        if (count > 0 && fcnt >= count) break;
-    }
-    return fcnt;
-}
-
-static int do_set_elem_get_sampling(htree_meta *htree, htree_node *node,
-                                    uint32_t remain, const uint32_t count,
-                                    set_elem_item **elem_array)
-{
-    int hidx;
-    int fcnt = 0; /* found count */
-
-    for (hidx = 0; hidx < HTREE_HASHTAB_SIZE; hidx++) {
-        if (node->hcnt[hidx] == -1) {
-            htree_node *child_node = (htree_node *)node->htab[hidx];
-            fcnt += do_set_elem_get_sampling(htree, child_node, remain,
-                                             count - fcnt, &elem_array[fcnt]);
-            remain -= child_node->tot_elem_cnt;
-        } else if (node->hcnt[hidx] > 0) {
-            set_elem_item *elem = node->htab[hidx];
-            while (elem != NULL) {
-                if ((rand() % remain) < (count - fcnt)) {
-                    elem->refcount++;
-                    elem_array[fcnt] = elem;
-                    fcnt++;
-                    if (fcnt >= count) break;
-                }
-                remain -= 1;
-                elem = elem->next;
-            }
-        }
-        if (fcnt >= count) break;
-    }
-    return fcnt;
-}
-
-static set_elem_item *do_set_elem_get_by_offset(htree_meta *htree, htree_node *node,
-                                                uint32_t offset, const bool delete,
-                                                size_t *space_decreased)
-{
-    int hidx;
-    for (hidx = 0; hidx < HTREE_HASHTAB_SIZE; hidx++) {
-        if (node->hcnt[hidx] == -1) {
-            htree_node *child_node = (htree_node *)node->htab[hidx];
-            if (offset >= child_node->tot_elem_cnt) {
-                offset -= child_node->tot_elem_cnt;
-                continue;
-            }
-            set_elem_item *found = do_set_elem_get_by_offset(htree, child_node, offset, delete, space_decreased);
-            if (delete) {
-                if (child_node->tot_elem_cnt < (HTREE_MAX_HASHCHAIN_SIZE/2)) {
-                    do_set_node_unlink(htree, node, hidx);
-                    *space_decreased += slabs_space_size(sizeof(htree_node));
-                }
-                node->tot_elem_cnt -= 1;
-            }
-            return found;
-        } else if (node->hcnt[hidx] > 0) {
-            if (offset >= node->hcnt[hidx]) {
-                offset -= node->hcnt[hidx];
-                continue;
-            }
-            set_elem_item *prev = NULL;
-            set_elem_item *elem = node->htab[hidx];
-            while (offset > 0) {
-                prev = elem;
-                elem = elem->next;
-                offset -= 1;
-            }
-            elem->refcount++;
-            if (delete) do_set_elem_unlink(node, hidx, prev, elem);
-            return elem;
-        }
-    }
-    return NULL;
-}
-
-static uint32_t do_set_elem_get_rand(htree_meta *htree,
-                                     const uint32_t ccnt,
-                                     const uint32_t count, const bool delete,
-                                     set_elem_item **elem_array,
-                                     size_t *space_decreased)
-{
-    assert(elem_array != NULL);
-    uint32_t fcnt = 0;
-
-    if (delete) { /* Deleting partial elements */
-        while (fcnt < count) {
-            int rand_offset = (rand() % ccnt);
-            set_elem_item *found = do_set_elem_get_by_offset(htree, htree->root,
-                                                             rand_offset, delete, space_decreased);
-            assert(found != NULL);
-            elem_array[fcnt++] = found;
-        }
-    } else if (count <= ccnt / 10) { /* Use hash table */
-        hash_table offset_ht;
-        if (!hash_init(&offset_ht, count))
-            return 0;
-        while (fcnt < count) {
-            int rand_offset = (rand() % ccnt);
-            if (hash_insert(&offset_ht, rand_offset)) {
-                set_elem_item *found = do_set_elem_get_by_offset(htree, htree->root,
-                                                                 rand_offset, false, NULL);
-                assert(found != NULL);
-                elem_array[fcnt++] = found;
-            }
-        }
-        hash_free(&offset_ht);
-    } else { /* Use sampling */
-        fcnt = do_set_elem_get_sampling(htree, htree->root, ccnt, count,
-                                        elem_array);
-        for (int i = fcnt - 1; i > 0; i--) {
-            int rand_idx = rand() % (i + 1);
-            if (rand_idx != i) {
-                set_elem_item *temp = elem_array[i];
-                elem_array[i] = elem_array[rand_idx];
-                elem_array[rand_idx] = temp;
-            }
-        }
-    }
-    return fcnt;
 }
 
 static uint32_t do_set_elem_get(set_meta_info *info,
@@ -560,17 +215,15 @@ static uint32_t do_set_elem_get(set_meta_info *info,
         CLOG_ELEM_DELETE_BEGIN((coll_meta_info*)info, count, cause);
     }
     if (count >= info->ccnt || count == 0) { /* Return all */
-        fcnt = do_set_elem_get_all(htree, htree->root, delete, elem_array, &space_decreased);
+        fcnt = htree_elem_get_bulk(htree, htree->root, 0 /* all */,
+                                   delete, (htree_elem_item **)elem_array, &space_decreased);
     } else { /* Return some */
-        fcnt = do_set_elem_get_rand(htree, info->ccnt, count, delete, elem_array, &space_decreased);
+        fcnt = htree_elem_get_rand(htree, info->ccnt, count,
+                                   delete, (htree_elem_item **)elem_array, &space_decreased);
     }
     if (delete) {
         for (int ii = 0; ii < fcnt; ii++) {
             do_set_elem_delete_post(info, elem_array[ii], cause);
-        }
-        if (htree->root->tot_elem_cnt == 0) {
-            do_set_node_unlink(htree, NULL, 0);
-            space_decreased += slabs_space_size(sizeof(htree_node));
         }
         if (info->stotal > 0 && space_decreased > 0) {
             do_coll_space_decr((coll_meta_info *)info, ITEM_TYPE_SET, space_decreased);
@@ -845,20 +498,16 @@ uint32_t set_elem_delete_with_count(set_meta_info *info, const uint32_t count)
 
     // cache_lock is held by caller
     if (htree->root != NULL) {
-        set_elem_item *deleted_head = NULL;
-        fcnt = do_set_elem_delete_by_count(htree, htree->root, count, &deleted_head, &space_decreased);
+        htree_elem_item *deleted_head = NULL;
+        fcnt = htree_elem_delete_bulk(htree, htree->root, count, &deleted_head, &space_decreased);
 
-        set_elem_item *deleted = deleted_head;
+        htree_elem_item *deleted = deleted_head;
         while (deleted != NULL) {
-            set_elem_item *next = deleted->next;
-            do_set_elem_delete_post(info, deleted, ELEM_DELETE_COLL);
+            htree_elem_item *next = deleted->next;
+            do_set_elem_delete_post(info, (set_elem_item *)deleted, ELEM_DELETE_COLL);
             deleted = next;
         }
 
-        if (htree->root->tot_elem_cnt == 0) {
-            do_set_node_unlink(htree, NULL, 0);
-            space_decreased += slabs_space_size(sizeof(htree_node));
-        }
         if (info->stotal > 0 && space_decreased > 0) { /* apply memory space */
             do_coll_space_decr((coll_meta_info *)info, ITEM_TYPE_SET, space_decreased);
         }
@@ -870,8 +519,8 @@ void set_elem_get_all(set_meta_info *info, elems_result_t *eresult)
 {
     htree_meta *htree = &info->htree;
     assert(eresult->elem_arrsz >= info->ccnt && eresult->elem_count == 0);
-    eresult->elem_count = do_set_elem_get_all(htree, htree->root, false,
-                                              (set_elem_item**)(eresult->elem_array), NULL);
+    eresult->elem_count = htree_elem_get_bulk(htree, htree->root, 0 /* all */,
+                                              false, (htree_elem_item**)(eresult->elem_array), NULL);
     assert(eresult->elem_count == info->ccnt);
 }
